API refactor

+++ monlixv2/src/main/java/com/monlixv2/service/models/Transaction.kt
@@ -4,7 +4,7 @@ data class Transaction(
     val id: String,
     val currency: String,
     val name: String,
-    val reward: Int,
+    val reward: Double,
     val status: String,
     val time: String
 )
